+++ tests/urls.py
@@ -29,19 +29,28 @@
 """Test project URL routing."""
 
 from django.conf.urls import include
-from django.conf.urls import url
 from django.http import HttpResponse
+from django.urls import re_path
 
 
 def blank_view(request):
-    """Return the hellow world view."""
+    """Return the hello world view."""
     return HttpResponse("<h1>HELLO WORLD!</h1>")
 
 
 urlpatterns = [
     # Include the allauth and 2FA urls from their respective packages.
-    url(r"^accounts/", include("allauth_2f2a.urls")),
-    url(r"^accounts/", include("allauth.urls")),
+    re_path(
+        r"^accounts/",
+        include("allauth_2f2a.urls"),
+    ),
+    re_path(
+        r"^accounts/",
+        include("allauth.urls"),
+    ),
     # A view without a name.
-    url(r"^unnamed-view$", blank_view),
+    re_path(
+        r"^unnamed-view$",
+        blank_view,
+    ),
 ]